Add inv admin.req
- Reform the tasks, separate utils out of tasks.
