+++ admin/tasks.py
@@ -1,9 +1,11 @@
 import os
 from invoke import task, run
+from tasks_utils import pip_install
 
 from website import settings
 
 HERE = os.path.dirname(os.path.abspath(__file__))
+WHEELHOUSE_PATH = os.environ.get('WHEELHOUSE')
 
 
 @task()
@@ -70,3 +72,9 @@ def bower_install():
     bower_bin = os.path.join(HERE, 'node_modules', 'bower', 'bin', 'bower')
     run('{} prune'.format(bower_bin), echo=True)
     run('{} install'.format(bower_bin), echo=True)
+
+
+@task(aliases=['req'])
+def requirements():
+    req_file = os.path.join(HERE, '..', 'requirements', 'admin.txt')
+    run(pip_install(req_file), echo=True)
